requirements/base.txt
# django>=1.7,<1.8
-e git+git@github.com:django/django.git@stable/1.7.x#egg=django

pillow>=2.4.0,<2.5

django-autoslug>=1.7.2,<1.8

pyisbn>=1.0.0,<1.1.0

djangorestframework>=2.3.13,<2.4.0
#-e git+git@github.com:9gix/django-rest-framework.git@unstable-merged#egg=djangorestframework

django-oauth-toolkit>=0.7.1,<0.8

python-social-auth>=0.1.26,<0.2

django-cors-headers>=0.12,<0.13

markdown>=2.4,<2.5

django-filter>=0.7,<0.8

gdata>=2.0.18,<2.1

# django-extensions>=1.3.3,<1.4
-e git+git@github.com:9gix/django-extensions.git@bugfix/printsql-django17#egg=django-extensions

bpython>=0.12,<1

ipython>=2.1.0,<2.2

pyzmq>=14.3.0,<14.4

celery>=3.1.9,<3.2

elasticsearch>=1.0.0,<1.1

pyelasticsearch>=0.6.1,<0.7

#django-haystack>=2.1.0,<2.2
-e git+git@github.com:toastdriven/django-haystack.git@master#egg=django-haystack

python-dateutil>=2.2,<2.3
